Added three files.
The former gui.py is now gui_pyqt5.py. This GUI was developed without Qt Designer.
The current gui.py that is uploaded is the GUI developed with Qt Designer.
mainwindow.ui is the user interface from Qt Designer.

I will continue to polish the GUI. My goal is to have the entire GUI completed in PyQt5 by Sprint #2.